[Issue139] Fix wrong data type for elixirLeaked

+++ src/main/java/jcrapi2/api/intern/players/battlelog/Opponent.java
@@ -43,7 +43,7 @@ public class Opponent {
   @SerializedName("cards")
   private List<Card> cards;
   @SerializedName("elixirLeaked")
-  private int elixirLeaked;
+  private double elixirLeaked;
   @SerializedName("rounds")
   private List<Round> rounds;
